Adds the ability to hide posts that have been labeled
