+++ .env.example
@@ -23,3 +23,12 @@ FEEDGEN_PUBLISHER_DID="did:plc:abcde...."
 
 # Delay between reconnect attempts to the firehose subscription endpoint (in milliseconds)
 FEEDGEN_SUBSCRIPTION_RECONNECT_DELAY=3000
+
+# Moderation services to subscribe to, comma-separated
+# N.b., the user handle might not be the hostname! Go to:
+# https://web.plc.directory/resolve
+# ...and find the #AtprotoLabeler record
+# LABELER_HOSTNAMES="mod.bsky.app"
+
+# Labels to hide on, comma-separated
+# HIDE_POSTS_WITH_LABELS="intolerant,misleading"
