+++ package.json
@@ -13,6 +13,8 @@
   },
   "dependencies": {
     "@atproto/api": "^0.6.20",
+    "@atproto/bsky": "^0.0.43",
+    "@atproto/common": "^0.4.0",
     "@atproto/identity": "^0.2.1",
     "@atproto/lexicon": "^0.2.2",
     "@atproto/repo": "^0.3.2",
