+++ src/config.ts
@@ -16,4 +16,6 @@ export type Config = {
   serviceDid: string
   publisherDid: string
   subscriptionReconnectDelay: number
+  labelerServices: string[]
+  moderationLabels: string[]
 }
