+++ src/server.ts
@@ -6,9 +6,10 @@ import { createServer } from './lexicon'
 import feedGeneration from './methods/feed-generation'
 import describeGenerator from './methods/describe-generator'
 import { createDb, Database, migrateToLatest } from './db'
-import { FirehoseSubscription } from './subscription'
+import { FirehoseSubscription } from './firehose-subscription'
 import { AppContext, Config } from './config'
 import wellKnown from './well-known'
+import { LabelerSubscription } from './labeler-subscription'
 
 export class FeedGenerator {
   public app: express.Application
@@ -16,7 +17,7 @@ export class FeedGenerator {
   public db: Database
   public firehose: FirehoseSubscription
   public cfg: Config
-
+  public labelers: LabelerSubscription[]
   constructor(
     app: express.Application,
     db: Database,
@@ -27,13 +28,15 @@ export class FeedGenerator {
     this.db = db
     this.firehose = firehose
     this.cfg = cfg
+    this.labelers = cfg.labelerServices.map(
+      (hostname) => new LabelerSubscription(db, hostname),
+    )
   }
 
   static create(cfg: Config) {
     const app = express()
     const db = createDb(cfg.sqliteLocation)
     const firehose = new FirehoseSubscription(db, cfg.subscriptionEndpoint)
-
     const didCache = new MemoryCache()
     const didResolver = new DidResolver({
       plcUrl: 'https://plc.directory',
@@ -64,6 +67,10 @@ export class FeedGenerator {
   async start(): Promise<http.Server> {
     await migrateToLatest(this.db)
     this.firehose.run(this.cfg.subscriptionReconnectDelay)
+    this.labelers.forEach((labelerService) =>
+      labelerService.run(this.cfg.subscriptionReconnectDelay),
+    )
+
     this.server = this.app.listen(this.cfg.port, this.cfg.listenhost)
     await events.once(this.server, 'listening')
     return this.server
